+++ Runtime/BlackboardReference.cs
@@ -41,11 +41,12 @@ namespace Unity.Behavior
         /// <param name="name">The name of the variable</param> 
         /// <param name="value">The value to assign to be assigned to the variable</param>
         /// <typeparam name="TValue">The type of value stored by the variable</typeparam>
+        /// <param name="logWarning">A bool to decide if a warning should be logged if adding fails</param>
         /// <returns>Returns true if the variable is added and false if a matching variable already exists.</returns>
-        public bool AddVariable<TValue>(string name, TValue value)
+        public bool AddVariable<TValue>(string name, TValue value, bool logWarning = true)
         {
             var wasSuccessful = Blackboard.AddVariable(name, value);
-            if (!wasSuccessful)
+            if (logWarning && !wasSuccessful)
             {
                 Debug.LogWarning($"Failed to add variable {name} to Blackboard.");
             }
@@ -60,11 +61,12 @@ namespace Unity.Behavior
         /// <param name="name">The name of the variable</param> 
         /// <param name="variable">The blackboard variable matching the name and value type</param>
         /// <typeparam name="TValue">The type of value stored by the variable</typeparam>
+        /// <param name="logWarning">A bool to decide if a warning should be logged when not found</param>
         /// <returns>Returns true if a variable matching the name and type is found. Returns false otherwise.</returns>
-        public bool GetVariable<TValue>(string name, out BlackboardVariable<TValue> variable)
+        public bool GetVariable<TValue>(string name, out BlackboardVariable<TValue> variable, bool logWarning = true)
         {
             var wasSuccessful = Blackboard.GetVariable(name, out variable);
-            if (!wasSuccessful)
+            if (logWarning && !wasSuccessful)
             {
                 Debug.LogWarning($"Failed to get variable {name} from Blackboard.");
             }
@@ -78,11 +80,12 @@ namespace Unity.Behavior
         /// <param name="name">The name of the variable</param> 
         /// <param name="variable">Contains the value associated with the specified name, if the named variable is found;
         /// otherwise, the default value is assigned.</param>
+        /// <param name="logWarning">A bool to decide if a warning should be logged when not found</param>
         /// <returns>Returns true if a variable matching the name and type is found. Returns false otherwise.</returns>
-        public bool GetVariable(string name, out BlackboardVariable variable)
+        public bool GetVariable(string name, out BlackboardVariable variable, bool logWarning = true)
         {
             var wasSuccessful = Blackboard.GetVariable(name, out variable);
-            if (!wasSuccessful)
+            if (logWarning && !wasSuccessful)
             {
                 Debug.LogWarning($"Failed to get variable {name} from Blackboard.");
             }
@@ -97,10 +100,10 @@ namespace Unity.Behavior
         /// otherwise, the default value is assigned.</param>
         /// <typeparam name="TValue">The type of value stored by the variable</typeparam>
         /// <returns>Returns true if a variable matching the name and type is found. Returns false otherwise.</returns>
-        public bool GetVariableValue<TValue>(string name, out TValue value)
+        public bool GetVariableValue<TValue>(string name, out TValue value, bool logWarning = true)
         {
             var wasSuccessful = Blackboard.GetVariableValue(name, out value);
-            if (!wasSuccessful)
+            if (logWarning && !wasSuccessful)
             {
                 Debug.LogWarning($"Failed to get variable {name}'s value from Blackboard.");
             }
@@ -113,11 +116,12 @@ namespace Unity.Behavior
         /// <param name="name">The name of the variable</param>
         /// <param name="value">The value to assign to the variable</param>
         /// <typeparam name="TValue">The type of value stored by the variable</typeparam>
+        /// <param name="logWarning">A bool to decide if a warning should be logged when not found</param>
         /// <returns>Returns true if a variable matching the name and type is found and set. Returns false otherwise.</returns>
-        public bool SetVariableValue<TValue>(string name, TValue value)
+        public bool SetVariableValue<TValue>(string name, TValue value, bool logWarning = true)
         {
             var wasSuccessful = Blackboard.SetVariableValue(name, value);
-            if (!wasSuccessful)
+            if (logWarning && !wasSuccessful)
             {
                 Debug.LogWarning($"Failed to set variable {name}'s value in the Blackboard.");
             }
@@ -130,27 +134,29 @@ namespace Unity.Behavior
         /// <param name="name">The name of the variable</param> 
         /// <param name="id">Contains the ID associated with the specified name, if the named variable is found;
         /// otherwise, the default value is assigned.</param>
+        /// <param name="logWarning">A bool to decide if a warning should be logged when not found</param>
         /// <returns>Returns true if a variable matching the name and type is found. Returns false otherwise.</returns>
-        public bool GetVariableID(string name, out SerializableGUID id)
+        public bool GetVariableID(string name, out SerializableGUID id, bool logWarning = true)
         {
             var wasSuccessful = Blackboard.GetVariableID(name, out id);
-            if (!wasSuccessful)
+            if (logWarning && !wasSuccessful)
             {
                 Debug.LogWarning($"Failed to get variable {name}'s id from the Blackboard.");
             }
             return wasSuccessful;
         }
-        
+
         /// <summary>
         /// Gets a variable associated with the specified GUID.
         /// </summary>
         /// <param name="guid">The GUID of the variable to get</param>
         /// <param name="variable">The variable associated with the specified GUID.</param>
+        /// <param name="logWarning">A bool to decide if a warning should be logged when not found</param>
         /// <returns>Returns true if a variable with a matching GUID was found and false otherwise.</returns>
-        public bool GetVariable(SerializableGUID guid, out BlackboardVariable variable)
+        public bool GetVariable(SerializableGUID guid, out BlackboardVariable variable, bool logWarning = true)
         {
             var wasSuccessful = Blackboard.GetVariable(guid, out variable);
-            if (!wasSuccessful)
+            if (logWarning && !wasSuccessful)
             {
                 Debug.LogWarning($"Failed to get variable with the guid {guid} from the Blackboard.");
             }
@@ -163,11 +169,12 @@ namespace Unity.Behavior
         /// <param name="guid">The GUID of the variable to get</param>
         /// <param name="variable">The variable associated with the specified GUID.</param>
         /// <typeparam name="TValue">The value type of the variable</typeparam>
+        /// <param name="logWarning">A bool to decide if a warning should be logged when not found</param>
         /// <returns>Returns true if a variable with a matching GUID and type was found and false otherwise.</returns>
-        public bool GetVariable<TValue>(SerializableGUID guid, out BlackboardVariable<TValue> variable)
+        public bool GetVariable<TValue>(SerializableGUID guid, out BlackboardVariable<TValue> variable, bool logWarning = true)
         {
             var wasSuccessful = Blackboard.GetVariable(guid, out variable);
-            if (!wasSuccessful)
+            if (logWarning && !wasSuccessful)
             {
                 Debug.LogWarning($"Failed to get variable with the guid {guid} from the Blackboard.");
             }
@@ -180,11 +187,12 @@ namespace Unity.Behavior
         /// <param name="guid">The guid associated with the variable</param>
         /// <param name="value">The value to assign to the variable</param>
         /// <typeparam name="TValue">The value type of the variable</typeparam>
+        /// <param name="logWarning">A bool to decide if a warning should be logged when not found</param>
         /// <returns>Returns true if the value was set successfully and false otherwise.</returns>
-        public bool SetVariableValue<TValue>(SerializableGUID guid, TValue value)
+        public bool SetVariableValue<TValue>(SerializableGUID guid, TValue value, bool logWarning = true)
         {
             var wasSuccessful = Blackboard.SetVariableValue(guid, value);
-            if (!wasSuccessful)
+            if (logWarning && !wasSuccessful)
             {
                 Debug.LogWarning($"Failed to set the variable with the guid {guid} in the Blackboard.");
             }