+++ Runtime/Execution/Conditions/Condition.cs
@@ -1,4 +1,6 @@
 using System;
+using Unity.Properties;
+using UnityEngine;
 
 namespace Unity.Behavior
 {
@@ -23,5 +25,16 @@ namespace Unity.Behavior
         /// OnEnd is called as a cleanup method for the Condition after it completes its logic check.
         /// </summary>
         public virtual void OnEnd() { }
+        
+        /// <summary>
+        /// The game object associated with the behavior graph.
+        /// </summary>
+        public GameObject GameObject => Graph.GameObject;
+        
+        /// <summary>
+        /// The BehaviorGraph containing the node instance.
+        /// </summary>
+        [SerializeReference]
+        internal BehaviorGraphModule Graph;
     }
 }